requirements.txt
fastapi==0.96.0
ipykernel==6.23.1
langchain==0.0.225
pypdf==3.8.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
uvicorn==0.22.0
torch==2.0.0
gradio==3.39.0
chromadb==0.4.3
openai==0.27.8
llama-cpp-python==0.1.78
